chore: extract auth utils
